requirements.txt
pydantic
requests
aiohttp
python-dotenv
bumpversion
pytest
pytest-asyncio
